requirements.txt
Django>=5.1
Pillow>=10.0.0
django-taggit>=4.0.0
sorl-thumbnail>=12.10.0

